homeChart&vehicle

+++ src/views/groupSetting/staffManage/positionRole.vue
@@ -35,15 +35,7 @@
       highlight-current-row
       style="width: 100%;min-height:500px;">
       <el-table-column
-        label="部门编码"
-        align="center"
-        width="120">
-        <template slot-scope="scope">
-          <span>{{ scope.row.id }}</span>
-        </template>
-      </el-table-column>
-      <el-table-column
-        label="岗位名称"
+        label="角色名称"
         width="150px"
         align="center">
         <template slot-scope="scope">
@@ -51,13 +43,13 @@
         </template>
       </el-table-column>
       <el-table-column
-        label="岗位描述"
+        label="角色描述"
         width="120px">
         <template slot-scope="scope">
           <span>{{ scope.row.description }}</span>
         </template>
       </el-table-column>
-      <el-table-column label="岗位职员">
+      <el-table-column label="角色标识">
         <template slot-scope="scope">
           <span>{{ scope.row.roleTag }}</span>
         </template>
@@ -107,12 +99,17 @@
         label-width="70px"
         style="width: 400px; margin-left:50px;">
         <el-form-item
-          label="岗位名称"
+          label="角色名称"
           prop="roleName">
           <el-input v-model="temp.roleName" />
         </el-form-item>
         <el-form-item
-          label="岗位描述"
+          label="角色标识"
+          prop="roleTag">
+          <el-input :disabled="dialogStatus === 'update'" v-model="temp.roleTag"/>
+        </el-form-item>
+        <el-form-item
+          label="角色描述"
           prop="description">
           <el-input v-model="temp.description" type="textarea"/>
         </el-form-item>
@@ -457,17 +454,29 @@ export default {
     },
     handleDelete(row) {
       // 在列表中删除 （将当前id传给后台）
-      deleteRole({ id: row.id }).then((response) => {
-        if (response.status === 200) {
-          if (response.data.code === 200) {
-            const index = this.list.indexOf(row)
-            this.list.splice(index, 1)
-            this.$notify({
-              title: '成功',
-              message: '删除成功',
-              type: 'success',
-              duration: 2000
-            })
+      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteRole({ id: row.id }).then((response) => {
+          if (response.status === 200) {
+            if (response.data.code === 200) {
+              const index = this.list.indexOf(row)
+              this.list.splice(index, 1)
+              this.$notify({
+                title: '成功',
+                message: '删除成功',
+                type: 'success',
+                duration: 2000
+              })
+            } else {
+              this.$notify.error({
+                title: '失败',
+                message: response.data.msg,
+                duration: 2000
+              })
+            }
           } else {
             this.$notify.error({
               title: '失败',
@@ -475,13 +484,12 @@ export default {
               duration: 2000
             })
           }
-        } else {
-          this.$notify.error({
-            title: '失败',
-            message: response.data.msg,
-            duration: 2000
-          })
-        }
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
     },
     handleCheck(data) {
